+++ src/Akeneo/Pim/Automation/DataQualityInsights/back/Domain/Model/Read/CriterionEvaluationCollection.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Akeneo\Pim\Automation\DataQualityInsights\Domain\Model\Read;
 
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\Axis;
+use Akeneo\Pim\Automation\DataQualityInsights\Domain\Model\ChannelLocaleRateCollection;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\CriterionCode;
 
 final class CriterionEvaluationCollection implements \IteratorAggregate, \Countable
@@ -36,6 +37,21 @@ final class CriterionEvaluationCollection implements \IteratorAggregate, \Counta
     public function get(CriterionCode $criterionCode): ?CriterionEvaluation
     {
         return $this->criteriaEvaluations[strval($criterionCode)] ?? null;
+    }
+
+    public function getCriterionRates(CriterionCode $criterionCode): ?ChannelLocaleRateCollection
+    {
+        $criterionEvaluation = $this->get($criterionCode);
+        if (null === $criterionEvaluation) {
+            return null;
+        }
+
+        $criterionEvaluationResult = $criterionEvaluation->getResult();
+        if (null === $criterionEvaluationResult) {
+            return null;
+        }
+
+        return $criterionEvaluationResult->getRates();
     }
 
     public function getIterator(): \Iterator