+++ src/Akeneo/Pim/Automation/DataQualityInsights/tests/back/Integration/Persistence/Query/GetLatestProductAxesRanksQueryIntegration.php
@@ -16,11 +16,15 @@ namespace Akeneo\Test\Pim\Automation\DataQualityInsights\Integration\Persistence
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\Model\Axis\Consistency;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\Model\Axis\Enrichment;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\Model\ChannelLocaleRankCollection;
+use Akeneo\Pim\Automation\DataQualityInsights\Domain\Model\ChannelLocaleRateCollection;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\Model\Read\AxisRankCollection;
+use Akeneo\Pim\Automation\DataQualityInsights\Domain\Model\Write\ProductAxisRates;
+use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\AxisCode;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\ChannelCode;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\LocaleCode;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\ProductId;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\Rank;
+use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\Rate;
 use Akeneo\Pim\Automation\DataQualityInsights\Infrastructure\Persistence\Query\GetLatestProductAxesRanksQuery;
 use Akeneo\Pim\Automation\DataQualityInsights\Infrastructure\Persistence\Repository\ProductAxisRateRepository;
 use Akeneo\Test\Integration\TestCase;
@@ -44,73 +48,59 @@ final class GetLatestProductAxesRanksQueryIntegration extends TestCase
 
     public function test_it_returns_the_latest_axes_rates_by_product_ids()
     {
+        $channelMobile = new ChannelCode('mobile');
+        $localeEn = new LocaleCode('en_US');
+        $localeFr = new LocaleCode('fr_FR');
+
         $rates = [
-            'product_42_consistency_latest_rates' => [
-                'product_id' => 42,
-                'axis' => 'consistency',
-                'evaluated_at' => new \DateTime('2020-01-08'),
-                'rates' => [
-                    'mobile' => [
-                        'en_US' => ['rank' => 1, 'value' => 96],
-                        'fr_FR' => ['rank' => 5, 'value' => 36],
-                    ]
-                ]
-            ],
-            'product_42_consistency_previous_rates' => [
-                'product_id' => 42,
-                'axis' => 'consistency',
-                'evaluated_at' => new \DateTime('2020-01-07'),
-                'rates' => [
-                    'mobile' => [
-                        'en_US' => ['rank' => 3, 'value' => 76],
-                        'fr_FR' => ['rank' => 4, 'value' => 67],
-                    ]
-                ]
-            ],
-            'product_42_enrichment_latest_rates' => [
-                'product_id' => 42,
-                'axis' => 'enrichment',
-                'evaluated_at' => new \DateTime('2020-01-08'),
-                'rates' => [
-                    'mobile' => [
-                        'en_US' => ['rank' => 2, 'value' => 84],
-                        'fr_FR' => ['rank' => 5, 'value' => 35],
-                    ]
-                ]
-            ],
-            'product_123_enrichment_latest_rates' => [
-                'product_id' => 123,
-                'axis' => 'enrichment',
-                'evaluated_at' => new \DateTime('2020-01-09'),
-                'rates' => [
-                    'mobile' => [
-                        'en_US' => ['rank' => 1, 'value' => 100],
-                        'fr_FR' => ['rank' => 1, 'value' => 95],
-                    ]
-                ]
-            ],
-            'product_123_enrichment_previous_rates' => [
-                'product_id' => 123,
-                'axis' => 'enrichment',
-                'evaluated_at' => new \DateTime('2020-01-08'),
-                'rates' => [
-                    'mobile' => [
-                        'en_US' => ['rank' => 2, 'value' => 81],
-                        'fr_FR' => ['rank' => 1, 'value' => 95],
-                    ]
-                ]
-            ],
-            'other_product_rates' => [
-                'product_id' => 456,
-                'axis' => 'enrichment',
-                'evaluated_at' => new \DateTime('2020-01-08'),
-                'rates' => [
-                    'mobile' => [
-                        'en_US' => ['rank' => 2, 'value' => 87],
-                        'fr_FR' => ['rank' => 1, 'value' => 95],
-                    ]
-                ]
-            ]
+            'product_42_consistency_latest_rates' => new ProductAxisRates(
+                new AxisCode('consistency'),
+                new ProductId(42),
+                new \DateTimeImmutable('2020-01-08'),
+                (new ChannelLocaleRateCollection())
+                    ->addRate($channelMobile, $localeEn, new Rate(96))
+                    ->addRate($channelMobile, $localeFr, new Rate(36))
+            ),
+            'product_42_consistency_previous_rates' => new ProductAxisRates(
+                new AxisCode('consistency'),
+                new ProductId(42),
+                new \DateTimeImmutable('2020-01-07'),
+                (new ChannelLocaleRateCollection())
+                    ->addRate($channelMobile, $localeEn, new Rate(76))
+                    ->addRate($channelMobile, $localeFr, new Rate(67))
+            ),
+            'product_42_enrichment_latest_rates' => new ProductAxisRates(
+                new AxisCode('enrichment'),
+                new ProductId(42),
+                new \DateTimeImmutable('2020-01-08'),
+                (new ChannelLocaleRateCollection())
+                    ->addRate($channelMobile, $localeEn, new Rate(84))
+                    ->addRate($channelMobile, $localeFr, new Rate(35))
+            ),
+            'product_123_enrichment_latest_rates' => new ProductAxisRates(
+                new AxisCode('enrichment'),
+                new ProductId(123),
+                new \DateTimeImmutable('2020-01-09'),
+                (new ChannelLocaleRateCollection())
+                    ->addRate($channelMobile, $localeEn, new Rate(100))
+                    ->addRate($channelMobile, $localeFr, new Rate(95))
+            ),
+            'product_123_enrichment_previous_rates' => new ProductAxisRates(
+                new AxisCode('enrichment'),
+                new ProductId(123),
+                new \DateTimeImmutable('2020-01-08'),
+                (new ChannelLocaleRateCollection())
+                    ->addRate($channelMobile, $localeEn, new Rate(81))
+                    ->addRate($channelMobile, $localeFr, new Rate(95))
+            ),
+            'other_product_rates' => new ProductAxisRates(
+                new AxisCode('enrichment'),
+                new ProductId(456),
+                new \DateTimeImmutable('2020-01-08'),
+                (new ChannelLocaleRateCollection())
+                    ->addRate($channelMobile, $localeEn, new Rate(87))
+                    ->addRate($channelMobile, $localeFr, new Rate(95))
+            ),
         ];
         $repository = $this->getRepository();
         $repository->save(array_values($rates));
@@ -124,17 +114,18 @@ final class GetLatestProductAxesRanksQueryIntegration extends TestCase
         $expectedRates = [
             42 => (new AxisRankCollection())
                 ->add($consistency->getCode(), (new ChannelLocaleRankCollection())
-                    ->addRank($channelMobile, $localeEn, Rank::fromInt($rates['product_42_consistency_latest_rates']['rates']['mobile']['en_US']['rank']))
-                    ->addRank($channelMobile, $localeFr, Rank::fromInt($rates['product_42_consistency_latest_rates']['rates']['mobile']['fr_FR']['rank']))
+                    ->addRank($channelMobile, $localeEn, Rank::fromRate($rates['product_42_consistency_latest_rates']->getRates()->getByChannelAndLocale($channelMobile, $localeEn)))
+                    ->addRank($channelMobile, $localeFr, Rank::fromRate($rates['product_42_consistency_latest_rates']->getRates()->getByChannelAndLocale($channelMobile, $localeFr)))
                 )
                 ->add($enrichment->getCode(), (new ChannelLocaleRankCollection())
-                    ->addRank($channelMobile, $localeEn, Rank::fromInt($rates['product_42_enrichment_latest_rates']['rates']['mobile']['en_US']['rank']))
-                    ->addRank($channelMobile, $localeFr, Rank::fromInt($rates['product_42_enrichment_latest_rates']['rates']['mobile']['fr_FR']['rank']))
+                    ->addRank($channelMobile, $localeEn, Rank::fromRate($rates['product_42_enrichment_latest_rates']->getRates()->getByChannelAndLocale($channelMobile, $localeEn)))
+                    ->addRank($channelMobile, $localeFr, Rank::fromRate($rates['product_42_enrichment_latest_rates']->getRates()->getByChannelAndLocale($channelMobile, $localeFr
+                    )))
                 ),
             123 => (new AxisRankCollection())
                 ->add($consistency->getCode(), (new ChannelLocaleRankCollection())
-                    ->addRank($channelMobile, $localeEn, Rank::fromInt($rates['product_123_enrichment_latest_rates']['rates']['mobile']['en_US']['rank']))
-                    ->addRank($channelMobile, $localeFr, Rank::fromInt($rates['product_123_enrichment_latest_rates']['rates']['mobile']['fr_FR']['rank']))
+                    ->addRank($channelMobile, $localeEn, Rank::fromRate($rates['product_123_enrichment_latest_rates']->getRates()->getByChannelAndLocale($channelMobile, $localeEn)))
+                    ->addRank($channelMobile, $localeFr, Rank::fromRate($rates['product_123_enrichment_latest_rates']->getRates()->getByChannelAndLocale($channelMobile, $localeFr)))
                 )
         ];